refactor(data_strucutres): move serialization helpers to separate file
