+++ data_structures/src/chain.rs
@@ -4,7 +4,7 @@ use secp256k1::{
     PublicKey as Secp256k1_PublicKey, SecretKey as Secp256k1_SecretKey,
     Signature as Secp256k1_Signature,
 };
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
+use serde::{Deserialize, Serialize};
 use std::{
     cell::Cell,
     collections::{BTreeMap, BTreeSet, HashMap, HashSet},
@@ -363,55 +363,6 @@ impl fmt::Debug for Hash {
     }
 }
 
-/// This is a copy of the Hash definition used to derive Serialize, Deserialize
-#[derive(Deserialize, Serialize)]
-enum HashSerializationHelper {
-    /// SHA-256 Hash
-    SHA256(SHA256),
-}
-
-impl From<Hash> for HashSerializationHelper {
-    fn from(x: Hash) -> Self {
-        match x {
-            Hash::SHA256(a) => HashSerializationHelper::SHA256(a),
-        }
-    }
-}
-
-impl Into<Hash> for HashSerializationHelper {
-    fn into(self) -> Hash {
-        match self {
-            HashSerializationHelper::SHA256(a) => Hash::SHA256(a),
-        }
-    }
-}
-
-impl Serialize for Hash {
-    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        if serializer.is_human_readable() {
-            serializer.collect_str(&self)
-        } else {
-            HashSerializationHelper::from(*self).serialize(serializer)
-        }
-    }
-}
-
-impl<'de> Deserialize<'de> for Hash {
-    fn deserialize<D>(deserializer: D) -> Result<Hash, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        use serde::de;
-        if deserializer.is_human_readable() {
-            String::deserialize(deserializer)?
-                .parse::<Hash>()
-                .map_err(de::Error::custom)
-        } else {
-            HashSerializationHelper::deserialize(deserializer).map(Into::into)
-        }
-    }
-}
-
 /// SHA-256 Hash
 pub type SHA256 = [u8; 32];
 
@@ -421,7 +372,7 @@ pub type SHA256 = [u8; 32];
 #[derive(Copy, Clone, Debug, Default, Eq, PartialEq, Hash, ProtobufConvert)]
 #[protobuf_convert(pb = "witnet::PublicKeyHash")]
 pub struct PublicKeyHash {
-    hash: [u8; 20],
+    pub(crate) hash: [u8; 20],
 }
 
 impl fmt::Display for PublicKeyHash {
@@ -465,50 +416,6 @@ impl PublicKeyHash {
         pkh.copy_from_slice(&h[..20]);
 
         Self { hash: pkh }
-    }
-}
-
-/// This is a copy of the Hash definition used to derive Serialize, Deserialize
-#[derive(Deserialize, Serialize)]
-struct PublicKeyHashSerializationHelper {
-    hash: [u8; 20],
-}
-
-impl From<PublicKeyHash> for PublicKeyHashSerializationHelper {
-    fn from(x: PublicKeyHash) -> Self {
-        Self { hash: x.hash }
-    }
-}
-
-impl Into<PublicKeyHash> for PublicKeyHashSerializationHelper {
-    fn into(self) -> PublicKeyHash {
-        PublicKeyHash { hash: self.hash }
-    }
-}
-
-impl Serialize for PublicKeyHash {
-    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        if serializer.is_human_readable() {
-            serializer.collect_str(&self)
-        } else {
-            PublicKeyHashSerializationHelper::from(*self).serialize(serializer)
-        }
-    }
-}
-
-impl<'de> Deserialize<'de> for PublicKeyHash {
-    fn deserialize<D>(deserializer: D) -> Result<PublicKeyHash, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        use serde::de;
-        if deserializer.is_human_readable() {
-            String::deserialize(deserializer)?
-                .parse::<PublicKeyHash>()
-                .map_err(de::Error::custom)
-        } else {
-            PublicKeyHashSerializationHelper::deserialize(deserializer).map(Into::into)
-        }
     }
 }
 
@@ -1107,57 +1014,6 @@ impl FromStr for OutputPointer {
             output_index,
             transaction_id,
         })
-    }
-}
-
-/// This is a copy of the Hash definition used to derive Serialize, Deserialize
-#[derive(Deserialize, Serialize)]
-struct OutputPointerSerializationHelper {
-    pub transaction_id: Hash,
-    pub output_index: u32,
-}
-
-impl From<OutputPointer> for OutputPointerSerializationHelper {
-    fn from(x: OutputPointer) -> Self {
-        OutputPointerSerializationHelper {
-            transaction_id: x.transaction_id,
-            output_index: x.output_index,
-        }
-    }
-}
-
-impl Into<OutputPointer> for OutputPointerSerializationHelper {
-    fn into(self) -> OutputPointer {
-        OutputPointer {
-            transaction_id: self.transaction_id,
-            output_index: self.output_index,
-        }
-    }
-}
-
-impl Serialize for OutputPointer {
-    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        if serializer.is_human_readable() {
-            serializer.collect_str(&self)
-        } else {
-            OutputPointerSerializationHelper::from(self.clone()).serialize(serializer)
-        }
-    }
-}
-
-impl<'de> Deserialize<'de> for OutputPointer {
-    fn deserialize<D>(deserializer: D) -> Result<OutputPointer, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        use serde::de;
-        if deserializer.is_human_readable() {
-            String::deserialize(deserializer)?
-                .parse::<OutputPointer>()
-                .map_err(de::Error::custom)
-        } else {
-            OutputPointerSerializationHelper::deserialize(deserializer).map(Into::into)
-        }
     }
 }
