+++ data_structures/src/lib.rs
@@ -24,5 +24,10 @@ pub mod error;
 /// Module containing data_request structures
 pub mod data_request;
 
+/// Serialization boilerplate to allow serializing some data structures as
+/// strings or bytes depending on the serializer.
+mod serialization_helpers;
+
+// TODO: tests should not be in `src/tests.rs`, there already exists a `tests/` folder for that
 #[cfg(test)]
 pub mod tests;
